We modify the CORS properties to add the HEAD.
In CategoryController, we just fix a bug we had on the endpoint.

ResourceController we prepare an endpoint that returns us a resource (UploadedFile) where we can download it in the frontend as a blob, and its securized with a header authorization.

Ticket controller we add the endpoints to: get all files from a ticket, add a file to the ticket, remove one file, remove all files from the ticket. We also got the remove status from a ticket (from the historic)

We modify the entity UploadedFile.

We create 2 new managers and 1 repository new.

Also we update the .gitignore so we can have in development a folder (uploads) to upload all the files