Solved little bug i had con CategoryManager.
We create an endpoint that tell us if we need to create a root user or not;

We syncronize file deleting so if for any chance 2 people delete the same file, does not give 400 to one of the users.

Added the get mapping for recive all users

We add one exception, and we modify 2 managers.

We also add the Resources class with static util methods